+++ packages/core/src/utils/chain-registry/chain-id-to-host-chain-name.ts
@@ -1,0 +1,5 @@
+const CHAIN_ID_SEPARATOR = '-'
+
+export const chainIdToHostChainName = (chainId: string) => {
+  return chainId.split(CHAIN_ID_SEPARATOR)[0]
+}