+++ packages/provider-cosmoskit/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @abstract-money/provider-cosmoskit
 
+## 5.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @abstract-money/react@1.3.7
+
 ## 5.0.6
 
 ### Patch Changes